FIX: Re Organize folders and files
