+++ public/src/entities/enemiesMenu.js
